# server/swagger.yaml
openapi: 3.0.0
info:
  title: GESTIONE_CERTIFICATI
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
  /login:
    post:
      summary: Login API
      security: []   # No security
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:      # Request body contents
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              example:   # Sample object
                username: admin
                password: 1a1dc91c907325c69271ddf0c944bc72
      responses:
        '200':
          description: OK
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /verifyToken:
    post:
      summary: Verify token API
      security: [] #No security
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /changePassword:
    post:
      summary: Change password API
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                passwordNew:
                  type: string
                passwordOld:
                  type: string
      responses:
        '200':
          description: OK
                

  #-----  User  -----
...
  /esami/:
    post:
      summary: CRUD ACTION create
      tags:
        - esame
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/esame"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nome:
                  type: string
    get:
      summary: CRUD ACTION list
      tags:
        - esame
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/esame"
  /esami/{id}:
    get:
      summary: CRUD ACTION get
      tags:
        - esame
      parameters:
        - name: id
          in: path
          description: Id 
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/esame"
    post:
      summary: CRUD ACTION update
      tags:
        - esame
      parameters:
        - name: id
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/esame"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/esame"
    
  #-----  voucher  -----
  /voucher/findByarea/{key}:
    get:
      summary: CRUD ACTION findByarea
      tags:
        - voucher
      parameters:
        - name: key
          in: path
          description: Id della risorsa area da cercare
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/voucher"
  /voucher/findBydipendente/{key}:
    get:
      summary: CRUD ACTION findBydipendente
      tags:
        - voucher
      parameters:
        - name: key
          in: path
          description: Id della risorsa dipendente da cercare
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/voucher"
  /voucher/findByesame/{key}:
    get:
      summary: CRUD ACTION findByesame
      tags:
        - voucher
      parameters:
        - name: key
          in: path
          description: Id della risorsa esame da cercare
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/voucher"
    
    
  #-----  User  -----
...
security:
  - bearerAuth: [] 
  
  
# Components
components:
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
...
    User:
      properties:
        _id:
          type: string
        mail:
          type: string
        name:
          type: string
        password:
          type: string
        roles:
          type: string
        surname:
          type: string
        username:
          type: string
      required:
        -  _id
        - password
        - username
...
    esame:
      properties:
        _id:
          type: string
        nome:
          type: string
      required:
        -  _id
        - nome
    voucher:
      properties:
        _id:
          type: string
        codiceVoucher:
          type: string
        dataAssegnazione:
          type: integer
        dataScadenza:
          type: integer
        note:
          type: string
        tipologiaAcquisto:
          type: string
      required:
        -  _id